Working implementation of Party Sheet. - Organized files from initial module. - Added party sheet to template.json (thinking we can store more in here). - Rewrote some code to get the implementation working. Mainly the sheet class was edited.
