--- src/party-sheet/.gitignore
@@ -1,3 +1,0 @@
-.idea/
-*.code-workspace
-.history
