--- src/party-sheet/README.md
@@ -1,38 +1,0 @@
-# Forbidden Lands Party Sheet
-A module for Foundry VTT for Forbidden Lands system.
-Adds a party sheet where you can visually assign party members to various travel activities. Provides convenient access to rolls required during travel.
-
-This module comes with convenient rules references for travel activities and rollable tables for mishaps and hunting (via compendium packs). These are provided with explicit permission from Free League Publishing.
-
-## Install
-1. Go to the setup page and choose **Modules**.
-2. Click the **Install Module** button, and paste in this [manifest link](https://raw.githubusercontent.com/maxstar/forbidden-lands-party-sheet/master/module.json)
-3. Launch your world and enable module in the settings.
-
-## How to use
-1. Create a new character and open its character sheet
-2. Go to sheet configuration ("Sheet" button in the sheet's title bar)
-3. In the first drop-down choose forbidden-lands.ForbiddenLandsPartySheet and press "Save Sheet Configuration" button
-4. GM must grant owner permission to all players for the party character
-5. Drag and drop player characters onto the party sheet
-6. Go to the "TRAVEL" tab and assign party members to various travel activities by drag&drop
-7. You can open journal entries with rules reference for a certain travel action by clicking on the travel activity name (GM must set proper permissions so that players are able to see these journal entries)
-8. There are buttons for all travel related rolls for your covenience
-9. Mishaps are NOT rolled automatically. GM has to manually make the roll.
-
-## Preview
-### Party Sheet
-![Party Sheet](https://github.com/maxstar/forbidden-lands-party-sheet/blob/master/assets/example.png?raw=true)
-
-## Translations
-- Czech by Mizy78
-- Russian by maxstar
-- Traditional Chinese by Rosataker
-
-## Related Websites
-- https://foundryvtt.com/
-- https://frialigan.se/en/store/?collection_id=84541866032
-
-## Licence
-Contents of rules references and rollable tables is owned by Free League Publishing and can only be distributed with their permission.
-Everything else is [GNU General Public License v3.0](https://choosealicense.com/licenses/gpl-3.0/)
