--- src/party-sheet/templates/partial/party-member.html
@@ -1,13 +1,0 @@
-<li class="party-member flexrow" data-entity-id="{{partyMember._id}}" draggable="true">
-    {{#if deleteButton }}
-        <div class="button">
-            <a class="item-control item-delete" title="Remove from party"><i class="fas fa-trash"></i></a>
-        </div>
-    {{/if}}
-    <img class="party-member-avatar" title="{{partyMember.name}}" src="{{partyMember.img}}">
-    {{#if noCharSheetLink }}
-        <span class="plain-name">{{partyMember.name}}</span>
-    {{else}}
-        {{ flps_enrich (flps_strconcat "@Actor[" partyMember.name "]") }}
-    {{/if}}
-</li>
