--- src/party-sheet/templates/partial/travel-action.html
@@ -1,22 +1,0 @@
-
-<div class="travel-action travel-action-{{action.key}}" data-travel-action="{{action.key}}">
-    <div class="travel-action-header">
-        {{#if action.journalEntryName}}
-            {{ flps_enrich (flps_strconcat "@JournalEntry[" (flps_capitalize action.journalEntryName) "]{" (localize action.name) "}") }}
-        {{else}}
-            {{ localize action.name }}
-        {{/if}}
-    </div>
-    <div class="travel-rolls">
-        {{#each action.buttons as |button|}}
-            <button class="travel-roll {{button.class}}">{{localize button.name}}</button>
-        {{/each}}
-    </div>
-    <div>
-        <ol>
-            {{#each assignedCharacters as |partyMember|}}
-                {{> modules/forbidden-lands-party-sheet/templates/partial/party-member.html partyMember=partyMember}}
-            {{/each}}
-        </ol>
-    </div>
-</div>
