--- src/party-sheet/templates/party.html
@@ -1,21 +1,0 @@
-<form class="{{cssClass}}" autocomplete="off">
-    <div class="party flex column">
-        <div class="sheet-tabs tabs flex row border" data-group="primary">
-            <b class="item" data-tab="main">{{localize "FLPS.TAB.MAIN"}}</b>
-            <b class="item" data-tab="travel">{{localize "FLPS.TAB.TRAVEL"}}</b>
-            <b class="item" data-tab="note">{{localize "TAB.NOTE"}}</b>
-        </div>
-
-        <div class="sheet-body flex row">
-            <div class="tab" data-group="primary" data-tab="main">
-                {{> modules/forbidden-lands-party-sheet/templates/tab/main.html}}
-            </div>
-            <div class="tab" data-group="primary" data-tab="travel">
-                {{> modules/forbidden-lands-party-sheet/templates/tab/travel.html}}
-            </div>
-            <div class="tab note" data-group="primary" data-tab="note">
-                {{editor content=data.bio.note.value target="data.bio.note.value" button=true owner=owner editable=editable}}
-            </div>
-        </div>
-    </div>
-</form>
