+++ src/styles/sheets/_index.scss
@@ -2,4 +2,5 @@
 @forward "item";
 @forward "monster";
 @forward "stronghold";
+@forward "party-sheet";
 @forward "tab";
