+++ src/templates/partial/travel-action.hbs
@@ -1,0 +1,20 @@
+<div class="travel-action travel-action-{{action.key}}" data-travel-action="{{action.key}}">
+	<div class="travel-action-header">
+		{{#if action.journalEntryName}} {{ flps_enrich (flps_strconcat "@JournalEntry[" (flps_capitalize
+		action.journalEntryName) "]{" (localize action.name) "}") }} {{else}} {{ localize action.name }} {{/if}}
+	</div>
+	<div class="travel-rolls">
+		{{#each action.buttons as |button|}}
+		<button class="travel-roll {{button.class}}">{{localize button.name}}</button>
+		{{/each}}
+	</div>
+	<div>
+		<ol>
+			{{#each assignedCharacters as |partyMember|}}
+			<!-- prettier-ignore -->
+			{{! prettier-ignore }}
+{{> systems/forbidden-lands/templates/partial/party-member.hbs partyMember=partyMember}}
+            {{/each}}
+		</ol>
+	</div>
+</div>
