+++ src/templates/tab/party-main.hbs
@@ -1,0 +1,28 @@
+<div class="flex column full">
+	<div class="flex row align-center border">
+		<div class="avatar">
+			<img src="{{actor.img}}" data-edit="img" title="{{actor.name}}" height="100" width="100" />
+		</div>
+		<div class="flex row margin align-end space-around grow">
+			<div class="flex column">
+				<div class="name">
+					<b>{{localize "FLPS.PARTY.NAME"}}</b>
+					<input name="name" type="text" value="{{actor.name}}" />
+				</div>
+			</div>
+		</div>
+	</div>
+	<div class="party-member-list border">
+		<h1>{{localize "FLPS.PARTY.MEMBERS" }}</h1>
+		<div class="tab-hint">{{localize "FLPS.PARTY.HINT_ADD"}}</div>
+		<div>
+			<ol>
+				{{#each partyMembers as |partyMember|}}
+				<!-- prettier-ignore -->
+				{{! prettier-ignore }}
+{{> systems/forbidden-lands/templates/partial/party-member.hbs partyMember=partyMember deleteButton=true}}
+				{{/each}}
+			</ol>
+		</div>
+	</div>
+</div>
